requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.1.0
pdfminer.six==20221105 